🚀 chore(platformio.ini): update LoRaMesher library dependency 🔧 chore(config.h): update MQTT_SERVER IP address and MQTT_MAX_PACKET_SIZE 🎉 feat(main.cpp): add support for Simulator service The LoRaMesher library dependency has been updated to the latest version. The MQTT_SERVER IP address has been updated to reflect the new IP address of the MQTT broker. The MQTT_MAX_PACKET_SIZE has been increased to 512 to allow for larger packets. The Simulator service has been added to the message manager to allow for simulation of messages.
🆕 feat(sim.cpp): add simulator functionality
This commit adds the simulator functionality to the codebase. The Sim class has been implemented with methods to start and stop the simulation, create a JSON object from a SimMessage, and process received messages. The simLoop method has been added to simulate the device's behavior by sending data every 15 minutes. The sendAllData method sends all the data in the statesList to the MQTT port. The createSimMessage method creates a SimMessage from an LM_State object.

🎉 feat(sim.h, simCommandService.cpp): add Sim class and SimCommandService class
The Sim class is a singleton class that extends the MessageService class. It contains methods to initialize the simulator, start and stop the simulator, and process received messages. The SimCommandService class is a class that contains commands to start and stop the simulator. It is used by the Sim class to execute the commands.

✨ feat(simCommandService.h): add SimCommandService class
A new header file `simCommandService.h` has been added, which defines a new class `SimCommandService`. This class inherits from the `CommandService` class and has a constructor. The `SimCommandService` class is used to simulate the behavior of the `CommandService` class.

🆕 feat(simMessage.h): add SimMessage class
The SimMessage class is added to the project. It is a derived class of DataMessageGeneric and contains a SimCommand enum and an LM_State object. The class has three functions: serialize, serializeState, and serializePacketHeader. The serialize function serializes the SimMessage object to a JSON object. The serializeState function serializes the LM_State object to a JSON object. The serializePacketHeader function serializes the packet header to a JSON object. The deserialize function deserializes the JSON object to a SimMessage object.